chore(context): reorganize and clean up tests
